requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-filter==24.2
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.7
mssql-django==1.5
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
